Made init into a ToolkitCommand so that it can use warn
